CBT-046 - Modify percentage data type from double to BigDecimal to avoid unexpected behaviors with decimal points

+++ src/main/java/com/distasilucas/cryptobalancetracker/model/response/crypto/CoinInfoResponse.java
@@ -7,7 +7,7 @@ public record CoinInfoResponse(
         String name,
         BigDecimal quantity,
         BigDecimal balance,
-        double percentage,
+        BigDecimal percentage,
         Set<String> platforms
 ) {
 }
